backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic[email]>=2.6.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.3
python-multipart>=0.0.9
phonenumbers>=9.0.25
cloudinary>=1.38.0
motor>=3.3.0